Rename `configurables.cpp` to `nsprefix.cpp` (#253)
