+++ NorthstarDLL/nsprefix.h
