+++ NorthstarDLL/serverauthentication.cpp
@@ -13,7 +13,7 @@
 #include <fstream>
 #include <filesystem>
 #include <thread>
-#include "configurables.h"
+#include "nsprefix.h"
 #include "nsmem.h"
 
 const char* AUTHSERVER_VERIFY_STRING = "I am a northstar server!";
